+++ api/org.integratedmodelling.klab.api/src/org/integratedmodelling/klab/api/runtime/ISession.java
@@ -19,6 +19,7 @@ import java.io.Closeable;
 import java.net.URL;
 import java.util.concurrent.Future;
 
+import org.integratedmodelling.klab.api.actors.IKlabActor;
 import org.integratedmodelling.klab.api.auth.IEngineSessionIdentity;
 import org.integratedmodelling.klab.api.auth.IEngineUserIdentity;
 import org.integratedmodelling.klab.api.auth.IUserIdentity;
@@ -49,11 +50,15 @@ import org.integratedmodelling.klab.exceptions.KlabException;
  * and its token must authenticate those engine API calls that are
  * session-aware. All sessions have a {@link IUserIdentity} as parent.
  * <p>
+ * If a session has a behavior associated (bound in the connection REST call by 
+ * name), it becomes an actor and implements it by setting priorities, views and
+ * whatever else the behavior specifies.
+ * <p>
  *
  * @author ferdinando.villa
  * @version $Id: $Id
  */
-public interface ISession extends IEngineSessionIdentity, Closeable {
+public interface ISession extends IKlabActor, IEngineSessionIdentity, Closeable {
 
 	/**
 	 * The observation action called on ISession always creates a new root subject.