+++ klab.tools/src/main/java/org/integratedmodelling/klab/clitool/console/commands/unit/Contextualize.java
@@ -7,7 +7,7 @@ import org.integratedmodelling.kim.api.IServiceCall;
 import org.integratedmodelling.klab.Units;
 import org.integratedmodelling.klab.api.cli.ICommand;
 import org.integratedmodelling.klab.api.data.mediation.IUnit;
-import org.integratedmodelling.klab.api.data.mediation.IUnit.Contextualization;
+import org.integratedmodelling.klab.api.data.mediation.IUnit.UnitContextualization;
 import org.integratedmodelling.klab.api.observations.scale.ExtentDimension;
 import org.integratedmodelling.klab.api.observations.scale.ExtentDistribution;
 import org.integratedmodelling.klab.api.runtime.ISession;
@@ -50,7 +50,7 @@ public class Contextualize implements ICommand {
             i++;
         }
 
-        Contextualization contextualization = Units.INSTANCE.getContextualization(unit, geometry, forcings);
+        UnitContextualization contextualization = Units.INSTANCE.getContextualization(unit, geometry, forcings);
         
         ret += "Chosen unit: " + contextualization.getChosenUnit();
         ret += "\nCandidates: ";